devtools/client/performance-new/test/chrome/test_perf-state-03.html

<!DOCTYPE HTML>
<html>
<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<head>
  <meta charset="utf-8">
  <title>Perf component test</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
  <div id="container"></div>

  <pre id="test">
    <script src="head.js" type="application/javascript"></script>
    <script type="application/javascript">
      "use strict";

      /**
       * Test the perf component for when the profiler is already started.
       */
      addPerfTest(async () => {
        const {
          perfFrontMock,
          getRecordingState,
          mountComponent,
          getState,
          selectors,
        } = createPerfComponent();

        mountComponent();

        is(getRecordingState(), "not-yet-known",
          "The component at first is in an unknown state.");

        await perfFrontMock._flushAsyncQueue();
        is(getRecordingState(), "available-to-record",
          "After talking to the actor, we're ready to record.");

        document.querySelector("button").click();
        is(getRecordingState(), "request-to-start-recording",
          "Sent in a request to start recording.");

        await perfFrontMock._flushAsyncQueue();
        is(getRecordingState(), "recording",
          "The actor has started its recording");

        ok(true, "Simulate a third party stopping the profiler.");
        ok(!selectors.getRecordingUnexpectedlyStopped(getState()),
          "The profiler has not unexpectedly stopped.");

        perfFrontMock.stopProfilerAndDiscardProfile();
        await perfFrontMock._flushAsyncQueue();

        ok(selectors.getRecordingUnexpectedlyStopped(getState()),
          "The profiler unexpectedly stopped.");
        is(getRecordingState(), "available-to-record",
          "However, the profiler is available to record again.");
      });
    </script>
  </pre>
</body>
</html>
